requirements.txt
django>=5.0
djangorestframework
python-dotenv
msal
boto3
openai
pillow
tqdm
geopy
sqlite-utils
requests>=2.31.0
pathlib2; python_version < "3.4"
